Just use the file length instead of the embedded length.
Code produced by CubePro embeds the file size in the first
couple of bytes of the file. However code from the cube-utils does
not. So use the actual file length to send to the printer so we can
print CubePro and our own encoded files.

Also the printer doesn't care about the file length in method 102
so don't send it, it makes things cleaner.